Roster: split off model code
